de-project-9/solution/service_dds/requirements.txt
APScheduler
confluent_kafka
flask
psycopg
psycopg-binary
pydantic
uuid
python-dotenv
